requirements.txt
fastapi
requests
uvicorn
redis
python-dotenv